# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Desk API
  version: 1.0.0
  description: |
    Multi-tenant office management platform API. Handles offices, leases, maintenance tickets,
    vendors, HVAC contracts, billing, and analytics.

    ## Authentication
    All protected endpoints require a JWT bearer token obtained from `/api/v1/auth/login`.
    API keys (prefixed `om_`) may be used as an alternative bearer credential.
servers:
  - url: /api/v1
    description: Primary API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or an API key prefixed with `om_`

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        pages: { type: integer }

    # --- Error --------------------------------------------------------------
...
    HvacContractResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        office_id: { type: string, format: uuid, nullable: true }
        vendor_id: { type: string, format: uuid, nullable: true }
        contract_number: { type: string, nullable: true }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        next_service_date: { type: string, format: date, nullable: true }
        annual_cost: { type: number, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        is_deleted: { type: boolean }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    HvacContractCreate:
      type: object
      properties:
        office_id: { type: string, format: uuid }
        vendor_id: { type: string, format: uuid }
        contract_number: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        next_service_date: { type: string, format: date }
        annual_cost: { type: number }
        status: { type: string }
        notes: { type: string }
    HvacContractUpdate:
      allOf:
        - $ref: '#/components/schemas/HvacContractCreate'
    PaginatedHvacContracts:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/HvacContractResponse' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # --- Notification -------------------------------------------------------
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /offices/{office_id}/hvac-contracts:
    get:
      tags: [Offices]
      summary: Get HVAC contracts for an office
      parameters:
        - { $ref: '#/components/parameters/OfficeId' }
      responses:
        '200':
          description: HVAC contracts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HvacContractResponse' }

  # ==========================================================================
  # LEASES
  # ==========================================================================
...
  /hvac-contracts:
    get:
      tags: [HvacContracts]
      summary: List HVAC contracts
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: office_id, in: query, schema: { type: string, format: uuid } }
        - { name: manager_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated HVAC contracts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedHvacContracts' }
    post:
      tags: [HvacContracts]
      summary: Create HVAC contract (admin/editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HvacContractCreate' }
      responses:
        '201':
          description: Created contract
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HvacContractResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /hvac-contracts/{contract_id}:
    put:
      tags: [HvacContracts]
      summary: Update HVAC contract (admin/editor)
      parameters:
        - { name: contract_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HvacContractUpdate' }
      responses:
        '200':
          description: Updated contract
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HvacContractResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [HvacContracts]
      summary: Soft-delete HVAC contract (admin only)
      parameters:
        - { name: contract_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Deleted }
  /hvac-contracts/{contract_id}/restore:
    patch:
      tags: [HvacContracts]
      summary: Restore soft-deleted HVAC contract (admin only)
      parameters:
        - { name: contract_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Restored contract
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HvacContractResponse' }

  # ==========================================================================
  # DASHBOARD
  # ==========================================================================
...